requirements.txt
openai
yfinance
feedparser
alpaca-py
python-dotenv
pandas
lxml
requests
